Update mining.js
Win system update

Add feature

+++ cmds/mining.js
@@ -6,11 +6,6 @@ const ms = require('ms')
 const moment = require("moment");
 require("moment-duration-format");
 
-function isOdd(num) {
-    if ((num % 5) == 0) return false;
-    else if ((num % 5) == 1) return true;
-}
-
 module.exports.run = async (bot, message, args) => {
 
     let bicon = message.author.displayAvatarURL;
@@ -28,11 +23,15 @@ module.exports.run = async (bot, message, args) => {
 
     let item = args[0];
     let moneyskull = "70000";
-    let moneygold = "100000";
-    let moneydiamond = "150000";
+    let moneygold = "120000";
+    let moneydiamond = "170000";
     let moneyrelic = "5000000";
     let reasons = ["Tu as croisé un fantôme tu as pris peur et quitté la mine en laissant tomber ton butin!", "Il y avait un panneau de radioactivité, tu as du rebrousser chemin!", "Tu es tombé dans une creuvase et tu as malheureusement du sacrifier ton butin pour te sauver la vie!", "Tu as cassé ta pioche !", "Tu as donné ton butin a l'orgresse pour lui faire un bisous tellement tu etait sous son charme !"];
     var reasonschoise = reasons[Math.floor(Math.random() * reasons.length)];
+    let danger = ["Sorcière(s)", "Squelette(s)", "Chauve-souris", "Démon(s)", "Gobelin(s)", "Ogresse(s)"]
+    var dangerchoise = danger[Math.floor(Math.random() * danger.length)];
+    let snowwhite = ["Te fais un bisous !", "Te fais un calin !", "Te fais la gueule !", "Te mouche le nez !", "T'essuie tes larmes !", "Te montre un téton !"]
+    var snowwhitechoise = snowwhite[Math.floor(Math.random() * snowwhite.length)];
     let pPick = pick[message.author.id].pick;
     let icon = message.author.displayAvatarURL;
 
@@ -85,9 +84,13 @@ module.exports.run = async (bot, message, args) => {
     console.log(`${message.author.username} est entré dans la mine a ${args[0]} | Temps estimé: ${time}`)
 
     setTimeout(function() {
-        let random = Math.floor(Math.random() * 35);
+        let relicrandom = Math.floor(Math.random() * 20);
+        let diamondrandom = Math.floor(Math.random() * 10);
+        let coinrandom = Math.floor(Math.random() * 8);
+        let skullrandom = Math.floor(Math.random() * 5);
+        let danger = Math.floor(Math.random() * 20);
 
-        if (random == 0 && item == 3) {
+        if (relicrandom == 0 && item == 3) {
             let curBal1 = coins[message.author.id].coins
             coins[message.author.id] = {
                 coins: curBal1 + parseInt(moneyrelic)
@@ -99,13 +102,14 @@ module.exports.run = async (bot, message, args) => {
                 .setTitle("**👑 Mine 👑**")
                 .addField("Butin trouvé dans la mine", args[0], true)
                 .addField("Tu remportes:", `${moneyrelic} Ecoin`, true)
+                .addField(`${dangerchoise} croisé:`, `${relicrandom}`)
                 .setImage("https://i.imgur.com/Oz6uHrH.gif")
                 .setFooter(`Butin de ${message.author.username}`, icon)
                 message.channel.send(embed);
                 message.delete().catch();
                 console.log(`${message.author.username} gagne ${moneyrelic} ecoin en ayant trouvé une ${args[0]} dans la mine`)
             });
-        } else if (isOdd(random) && item == 2) {
+        } else if (diamondrandom == 5 && item == 2) {
             let curBal2 = coins[message.author.id].coins
             coins[message.author.id] = {
                 coins: curBal2 + parseInt(moneydiamond)
@@ -117,13 +121,14 @@ module.exports.run = async (bot, message, args) => {
                 .setTitle("**:gem: Mine :gem:**")
                 .addField("Butin trouvé dans la mine:", args[0], true)
                 .addField("Tu remportes:", `${moneydiamond} Ecoin`, true)
+                .addField(`${dangerchoise} croisé:`, `${diamondrandom}`)
                 .setImage("https://i.imgur.com/Oz6uHrH.gif")
                 .setFooter(`Butin de ${message.author.username}`, icon)
                 message.channel.send(embed);
                 message.delete().catch();
                 console.log(`${message.author.username} gagne ${moneydiamond} ecoin en ayant trouvé un ${args[0]} dans la mine`)
             });
-        } else if (isOdd(random) && item == 1) {
+        } else if (coinrandom == 4 && item == 1) {
             let curBal3 = coins[message.author.id].coins
             coins[message.author.id] = {
                 coins: curBal3 + parseInt(moneygold)
@@ -135,13 +140,14 @@ module.exports.run = async (bot, message, args) => {
                 .setTitle("**💰 Mine 💰**")
                 .addField("Butin trouvé dans la mine:", args[0], true)
                 .addField("Tu remportes:", `${moneygold} Ecoin`, true)
+                .addField(`${dangerchoise} croisé:`, `${coinrandom}`)
                 .setImage("https://i.imgur.com/Oz6uHrH.gif")
                 .setFooter(`Butin de ${message.author.username}`, icon)
                 message.channel.send(embed);
                 message.delete().catch();
                 console.log(`${message.author.username} gagne ${moneygold} ecoin en ayant trouvé un ${args[0]} dans la mine`)
             });
-        } else if (!isOdd(random) && item == 0) {
+        } else if (skullrandom == 2 && item == 0) {
             let curBal3 = coins[message.author.id].coins
             coins[message.author.id] = {
                 coins: curBal3 + parseInt(moneyskull)
@@ -153,6 +159,7 @@ module.exports.run = async (bot, message, args) => {
                 .setTitle("**💀 Mine 💀**")
                 .addField("Butin trouvé dans la mine:", args[0], true)
                 .addField("Tu remportes:", `${moneyskull} Ecoin`, true)
+                .addField(`${dangerchoise} croisé:`, `${skullrandom}`)
                 .setImage("https://i.imgur.com/Oz6uHrH.gif")
                 .setFooter(`Butin de ${message.author.username}`, icon)
                 message.channel.send(embed);
@@ -162,14 +169,22 @@ module.exports.run = async (bot, message, args) => {
         } else {
             let embed = new Discord.RichEmbed()
             .setColor("#76685E")
-            .setTitle("**Tu as quitté la mine**")
+            .setTitle("**Pas de butin**")
             .addField("raison:", reasonschoise)
             .addField("butin perdu:", args[0], true)
+            .addField(`${dangerchoise} croisé:`, `${danger}`, true)
+            .addField(`Blanche neige:`, `${snowwhitechoise}`)
             .setImage("https://i.imgur.com/RSHfwrL.gif")
             .setFooter(`${message.author.username} a perdu son butin`, icon)
             message.channel.send(embed);
-            message.delete().catch();
+            let whitesnowembed = new Discord.RichEmbed()
+            .setTitle("**Blanche Neige**")
+            .setColor("#E642AA")
+            .setImage("https://i.imgur.com/aCnYiGg.jpg")
+            .setFooter(`${message.author.username}`, bicon);
+            if (snowwhitechoise === "Te montre un téton !") return message.author.send(whitesnowembed), console.log(`Blanche neige a montré un téton a ${message.author.username}`) , message.delete().catch();
             console.log(`${message.author.username} as quittés la mine pour la raison suivante: ${reasonschoise} et perd son butin ${args[0]} dans la mine`)
+            message.delete().catch();
         }
 
         console.log(`Commande !mining executé sur le serveur ${message.guild.name} dans le salon ${message.channel.name} par le membre ${message.author.username}`)
